+++ resources/templates/ts/retained.ts
@@ -4,7 +4,7 @@
 
 import * as vscode from 'vscode';
 
-// The macro context isn’t disposed when the script finishes; it must be
+// The macro context isn't disposed when the script finishes; it must be
 // explicitly stopped via the "Macros: Show Running Macros" command or the
 //  "Request to Stop" action in the Macro Explorer.
 
@@ -13,7 +13,7 @@ __disposables.push(
     if (editor) {
       vscode.window.showInformationMessage(
         `Active editor: ${vscode.workspace.asRelativePath(editor.document.uri)}. ` +
-          `Macro ${__runId} must be manually stopped for this listener to be disposed.`,
+        `Macro ${__runId} must be manually stopped for this listener to be disposed.`,
       );
     }
   }),